test(mlx): move Moonlight EOS regression test into a non-slow module
Addresses review on #305:
- The test was in test_tokenizers.py which carries module-level
  `pytestmark = pytest.mark.slow`; CI runs `pytest -m "not slow"`, so it was
  silently deselected and protected nothing. Move it to a dedicated non-slow
  test_eos_token_ids.py (pure mapping, no tokenizer download) so it runs in CI.
- Wrap model-id string literals in ModelId(...) to satisfy strict basedpyright
  (ModelId is a NewType). Also adds pure-mapping coverage for the other
  explicit-EOS families and the None (use-tokenizer-config) path.
